Besides fixing a few warnings and improving the performance test a bit, this commit changes the chunked allocation behavior.  When creating mesh through the ReadUtil interface, the previous code allocated at the default chunk size, defined in SequenceManager.  This presents a problem when representing a small mesh with many, large dense tags, since dense tags are allocated according to the chunk size.  This has been changed such that applications can specify a chunk size to the ReadUtil functions for creating nodes/elements.  The default behavior was changed to use the exact number of nodes/elements for the chunk size (or, when building in parallel, 1.5 times that number, to account for ghosts that get added later).  Applications calling the MOAB API functions (not the ReadUtil functions) for creating vertices/entities still allocate at the original default chunk sizes.
Passed make check in serial and parallel.

Specific changes:
SpectralMeshTool.cpp: remove a warning
ReadUtilIface, ReadUtil: add sequence size arguments to get_node_coords and get_element_connect,
  defaulted to -1 (which results in num_nodes/num_elements being used for sequence size)
SequenceManager: add a sequence size to create_entity_sequence, and change the way the
  sequence size is determined.  Also moved DEFAULT_*_SEQUENCE_SIZE variables to static public
  members, for use by other readers wanting to use the default size.
Core.cpp: initialized variable to remove warning.
Tqdcfr.cpp: explicitly specify default sequence sizes, since nodes/elements are allocated by
  geometric entity.  Also fixed a few compile warnings.
ReadTemplate.cpp, ReadGCRM.cpp: empty statements added to remove compiler warning
MOAB-UG.doc: added some info on chunked allocation
perf.cpp, runtest: changes from Gabriel Marin (ORNL) to allow calling iterator-based tests separately
test/MBTest.cpp: added sequence size parameter
WriteCCMIO.cpp: substituting the new tag_get_handle for a deprecated one.
FBEngine.cpp: getting rid of a few unused variables.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5898 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ src/SpectralMeshTool.cpp
@@ -125,7 +125,7 @@ ErrorCode SpectralMeshTool::create_spectral_elems(const T *conn, int num_fine_el
   if (MB_SUCCESS != rval) return rval;
   
   int verts_per_felem = spectralOrderp1*spectralOrderp1,
-      verts_per_celem = std::pow(2, dim);
+      verts_per_celem = std::pow((double)2.0, dim);
 
   rval = rmi->get_element_connect(num_coarse_elems, verts_per_celem,
                                   (2 == dim ? MBQUAD : MBHEX), 0, start_elem, new_conn);
